+++ CHANGELOG.md
@@ -50,3 +50,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.1] - 2023-05-25
 ### Added
 - option to filter for uncomplete entries
+
+## [0.1.2] - 2023-05-30
+### Added
+- add custom markups to the context menu of textfields
+- custom markups can now be configured inside the settings